Check input image_path

This function checks that the image_path vector has the appropriate structure. For more details please check the vignette documentation.

check_image_path(image_path, sce)

Arguments

image_path
A path to the directory containing the low resolution histology images that is needed for the interactive visualizations with plotly. See https://github.com/LieberInstitute/spatialLIBD/tree/master/inst/app/www/data for an example of how these files should be organized.

sce
Defaults to the output of fetch_data(type = 'sce'). This is a SingleCellExperiment object with the spot-level Visium data and information required for visualizing the histology. See fetch_data() for more details.

Value

The input object if all checks are passed.

See also

Examples

if (enough_ram()) {
    ## Obtain the necessary data
    if (!exists("sce")) sce <- fetch_data("sce")

    ## Get the path to the images
    img_path <- system.file("app", "www", "data", package = "spatialLIBD")

    ## Check the object
    check_image_path(img_path, sce)
}

#> using temporary cache /tmp/RtmpzGdRMO/BiocFileCache

#> snapshotDate(): 2020-09-14

#> adding rname 'https://www.dropbox.com/s/f4wcvtdq428y73p/Human_DLPFC_Visium_processedData_sce_scran_spatialLIBD.Rdata?dl=1'

#> Loading objects:
#>   sce

#> [1] "/usr/local/lib/R/host-site-library/spatialLIBD/app/www/data"