Check input modeling_results

This function checks that the modeling_results object has the appropriate structure. For more details please check the vignette documentation.

check_modeling_results(modeling_results)

Arguments

modeling_results

Defaults to the output of fetch_data(type = 'modeling_results'). This is a list of tables with the columns f_stat_* or t_stat_* as well as p_value_* and fdr_* plus ensembl. The column name is used to extract the statistic results, the p-values, and the FDR adjusted p-values. Then the ensembl column is used for matching in some cases. See fetch_data() for more details.

Value

The input object if all checks are passed.

See also

Examples

if (!exists("modeling_results")) {
    modeling_results <- fetch_data(type = "modeling_results")
}
#> 2024-07-17 13:49:14.441658 loading file /github/home/.cache/R/BiocFileCache/4981423d30a_Human_DLPFC_Visium_modeling_results.Rdata%3Fdl%3D1

## Check the object
xx <- check_modeling_results(modeling_results)