Import cluster results

This function imports previously exported clustering results with cluster_export() and adds them to the colData() slot of your SpatialExperiment-class object.

cluster_import(
  spe,
  cluster_dir = file.path(tempdir(), "exported_clusters"),
  prefix = "imported_",
  overwrite = TRUE
)

Arguments

spe

Defaults to the output of fetch_data(type = 'spe'). This is a SpatialExperiment-class object with the spot-level Visium data and information required for visualizing the histology. See fetch_data() for more details.

cluster_dir

A character(1) specifying the output directory, similar to the outs/analysis/clustering produced by SpaceRanger.

prefix

A character(1) specifying the prefix to use when naming these new cluster variables.

overwrite

A logical(1) indicating whether to overwrite the spe$key.

Value

A SpatialExperiment-class object with the imported clusters appended on the colData().

See also

Examples

if (enough_ram()) {
    ## Obtain the necessary data
    if (!exists("spe")) spe <- fetch_data("spe")

    ## Export two cluster variables
    cluster_export(spe, "spatialLIBD")
    cluster_export(spe, "GraphBased")

    ## Re-import them
    colData(cluster_import(spe))
}
#> snapshotDate(): 2024-04-29
#> 2024-07-11 14:08:32.632255 loading file /Users/leocollado/Library/Caches/org.R-project.R/R/BiocFileCache/6f56656adcb_Human_DLPFC_Visium_processedData_sce_scran_spatialLIBD.Rdata%3Fdl%3D1
#> Overwriting 'spe$key'. Set 'overwrite = FALSE' if you do not want to overwrite it.
#> Warning: '/var/folders/_2/hqcmhc1d1bl3lqxgcrkpv8_00000gn/T//RtmpCGZyme/exported_clusters/spatialLIBD' already exists
#> Overwriting 'spe$key'. Set 'overwrite = FALSE' if you do not want to overwrite it.
#> Warning: '/var/folders/_2/hqcmhc1d1bl3lqxgcrkpv8_00000gn/T//RtmpCGZyme/exported_clusters/GraphBased' already exists
#> Overwriting 'spe$key'. Set 'overwrite = FALSE' if you do not want to overwrite it.
#> DataFrame with 47681 rows and 71 columns
#>                                       key   sample_id   Cluster   sum_umi  sum_gene     subject
#>                               <character> <character> <integer> <numeric> <integer> <character>
#> AAACAACGAATAGTTC-1 AAACAACGAATAGTTC-1_1..      151507         6       948       727      Br5292
#> AAACAAGTATCTCCCA-1 AAACAAGTATCTCCCA-1_1..      151507         3      4261      2170      Br5292
#> AAACAATCTACTAGCA-1 AAACAATCTACTAGCA-1_1..      151507         2      1969      1093      Br5292
#> AAACACCAATAACTGC-1 AAACACCAATAACTGC-1_1..      151507         5      3368      1896      Br5292
#> AAACAGCTTTCAGAAG-1 AAACAGCTTTCAGAAG-1_1..      151507         1      2981      1620      Br5292
#> ...                                   ...         ...       ...       ...       ...         ...
#> TTGTTGTGTGTCAAGA-1 TTGTTGTGTGTCAAGA-1_1..      151676         5      2880      1638      Br8100
#> TTGTTTCACATCCAGG-1 TTGTTTCACATCCAGG-1_1..      151676         2      2640      1526      Br8100
#> TTGTTTCATTAGTCTA-1 TTGTTTCATTAGTCTA-1_1..      151676         2      2063      1279      Br8100
#> TTGTTTCCATACAACT-1 TTGTTTCCATACAACT-1_1..      151676         7      1031       748      Br8100
#> TTGTTTGTGTAAATTC-1 TTGTTTGTGTAAATTC-1_1..      151676         1      2625      1392      Br8100
#>                       position   replicate subject_position   discard cell_count SNN_k50_k4
#>                    <character> <character>      <character> <logical>  <integer>  <integer>
#> AAACAACGAATAGTTC-1           0           1      Br5292_pos0     FALSE          0          1
#> AAACAAGTATCTCCCA-1           0           1      Br5292_pos0     FALSE          3          1
#> AAACAATCTACTAGCA-1           0           1      Br5292_pos0     FALSE          2          3
#> AAACACCAATAACTGC-1           0           1      Br5292_pos0     FALSE          5          2
#> AAACAGCTTTCAGAAG-1           0           1      Br5292_pos0     FALSE          3          1
#> ...                        ...         ...              ...       ...        ...        ...
#> TTGTTGTGTGTCAAGA-1         300           2    Br8100_pos300     FALSE          1          1
#> TTGTTTCACATCCAGG-1         300           2    Br8100_pos300     FALSE          6          2
#> TTGTTTCATTAGTCTA-1         300           2    Br8100_pos300     FALSE          9          2
#> TTGTTTCCATACAACT-1         300           2    Br8100_pos300     FALSE          2          1
#> TTGTTTGTGTAAATTC-1         300           2    Br8100_pos300     FALSE          0          1
#>                    SNN_k50_k5 SNN_k50_k6 SNN_k50_k7 SNN_k50_k8 SNN_k50_k9 SNN_k50_k10 SNN_k50_k11
#>                     <integer>  <integer>  <integer>  <integer>  <integer>   <integer>   <integer>
#> AAACAACGAATAGTTC-1          1          1          3          3          4           4           5
#> AAACAAGTATCTCCCA-1          1          1          1          1          1           1           4
#> AAACAATCTACTAGCA-1          2          3          4          4          5           5           6
#> AAACACCAATAACTGC-1          3          4          5          8          9          10          11
#> AAACAGCTTTCAGAAG-1          1          1          1          1          3           3           3
#> ...                       ...        ...        ...        ...        ...         ...         ...
#> TTGTTGTGTGTCAAGA-1          1          1          1          1          1           1           4
#> TTGTTTCACATCCAGG-1          3          4          5          6          7           8           9
#> TTGTTTCATTAGTCTA-1          3          4          5          6          7           8           9
#> TTGTTTCCATACAACT-1          1          1          1          1          3           3           3
#> TTGTTTGTGTAAATTC-1          1          2          2          2          2           2           1
#>                    SNN_k50_k12 SNN_k50_k13 SNN_k50_k14 SNN_k50_k15 SNN_k50_k16 SNN_k50_k17
#>                      <integer>   <integer>   <integer>   <integer>   <integer>   <integer>
#> AAACAACGAATAGTTC-1           5           5           6           7           7           9
#> AAACAAGTATCTCCCA-1           4           4           4           3           3           3
#> AAACAATCTACTAGCA-1           6           6           7           8           8           7
#> AAACACCAATAACTGC-1          10           9          10          11          11          12
#> AAACAGCTTTCAGAAG-1           3           3           3           2           2           2
#> ...                        ...         ...         ...         ...         ...         ...
#> TTGTTGTGTGTCAAGA-1           4           4           4           3           3           3
#> TTGTTTCACATCCAGG-1           8           8           9          10          10          11
#> TTGTTTCATTAGTCTA-1           8           8           9          10          10          11
#> TTGTTTCCATACAACT-1           3           3           3           2           2           2
#> TTGTTTGTGTAAATTC-1           1           1           1           5           5           5
#>                    SNN_k50_k18 SNN_k50_k19 SNN_k50_k20 SNN_k50_k21 SNN_k50_k22 SNN_k50_k23
#>                      <integer>   <integer>   <integer>   <integer>   <integer>   <integer>
#> AAACAACGAATAGTTC-1          10          10          10          10          11          11
#> AAACAAGTATCTCCCA-1           2           2           2           2           1           1
#> AAACAATCTACTAGCA-1           8           8           8           8           9           9
#> AAACACCAATAACTGC-1          13          12          11          11          12          20
#> AAACAGCTTTCAGAAG-1           1           1           1           1           6           6
#> ...                        ...         ...         ...         ...         ...         ...
#> TTGTTGTGTGTCAAGA-1           2           2           2           2           1           1
#> TTGTTTCACATCCAGG-1          12          18          18          17          18          17
#> TTGTTTCATTAGTCTA-1          12          15          15          14          15          14
#> TTGTTTCCATACAACT-1           1           1           1           1           6           6
#> TTGTTTGTGTAAATTC-1           5           5           5           5           4           4
#>                    SNN_k50_k24 SNN_k50_k25 SNN_k50_k26 SNN_k50_k27 SNN_k50_k28 GraphBased  Maynard
#>                      <integer>   <integer>   <integer>   <integer>   <integer>  <integer> <factor>
#> AAACAACGAATAGTTC-1          11          12          13          13          14          6       WM
#> AAACAAGTATCTCCCA-1           1           6           5           5           4          3       6 
#> AAACAATCTACTAGCA-1           9           9           9           9           9          2       1 
#> AAACACCAATAACTGC-1          20          21          22          21          22          5       5 
#> AAACAGCTTTCAGAAG-1           6           5           4           4           3          1       4 
#> ...                        ...         ...         ...         ...         ...        ...      ...
#> TTGTTGTGTGTCAAGA-1           1          10          10          10          10          5       5 
#> TTGTTTCACATCCAGG-1          17          18          19          18          19          2       WM
#> TTGTTTCATTAGTCTA-1          13          14          15          15          16          2       WM
#> TTGTTTCCATACAACT-1           6           5           4           4           3          7       6 
#> TTGTTTGTGTAAATTC-1           4           3           3           3           2          1       1 
#>                    Martinowich layer_guess layer_guess_reordered layer_guess_reordered_short
#>                       <factor>    <factor>              <factor>                    <factor>
#> AAACAACGAATAGTTC-1          WM      Layer1                Layer1                          L1
#> AAACAAGTATCTCCCA-1          6       Layer3                Layer3                          L3
#> AAACAATCTACTAGCA-1          1       Layer1                Layer1                          L1
#> AAACACCAATAACTGC-1          5       WM                    WM                              WM
#> AAACAGCTTTCAGAAG-1          4       Layer6                Layer6                          L6
#> ...                        ...         ...                   ...                         ...
#> TTGTTGTGTGTCAAGA-1          4       Layer6                Layer6                          L6
#> TTGTTTCACATCCAGG-1          WM      WM                    WM                              WM
#> TTGTTTCATTAGTCTA-1          WM      WM                    WM                              WM
#> TTGTTTCCATACAACT-1          6       Layer6                Layer6                          L6
#> TTGTTTGTGTAAATTC-1          1       Layer1                Layer1                          L1
#>                    expr_chrM expr_chrM_ratio SpatialDE_PCA SpatialDE_pool_PCA   HVG_PCA
#>                    <numeric>       <numeric>     <integer>          <integer> <integer>
#> AAACAACGAATAGTTC-1        88       0.0928270             4                  5         3
#> AAACAAGTATCTCCCA-1       657       0.1541892             2                  3         1
#> AAACAATCTACTAGCA-1       431       0.2188928             4                  4         4
#> AAACACCAATAACTGC-1       272       0.0807601             6                  7         6
#> AAACAGCTTTCAGAAG-1       502       0.1683999             1                  1         3
#> ...                      ...             ...           ...                ...       ...
#> TTGTTGTGTGTCAAGA-1       475        0.164931             1                  2         2
#> TTGTTTCACATCCAGG-1       320        0.121212             5                  5         4
#> TTGTTTCATTAGTCTA-1       262        0.127000             5                  5         4
#> TTGTTTCCATACAACT-1       161        0.156159             4                  8         6
#> TTGTTTGTGTAAATTC-1       667        0.254095             1                  1         2
#>                    pseudobulk_PCA markers_PCA SpatialDE_UMAP SpatialDE_pool_UMAP  HVG_UMAP
#>                         <integer>   <integer>      <integer>           <integer> <integer>
#> AAACAACGAATAGTTC-1              5           4              1                   2         1
#> AAACAAGTATCTCCCA-1              2           2              1                   3         1
#> AAACAATCTACTAGCA-1              4           8              2                   4         5
#> AAACACCAATAACTGC-1              6           8              5                   7         7
#> AAACAGCTTTCAGAAG-1              1           1              2                   3         6
#> ...                           ...         ...            ...                 ...       ...
#> TTGTTGTGTGTCAAGA-1              2           1              1                   3         2
#> TTGTTTCACATCCAGG-1              5           6              4                   5         5
#> TTGTTTCATTAGTCTA-1              4           1              4                   5         7
#> TTGTTTCCATACAACT-1              2           3              1                   2         3
#> TTGTTTGTGTAAATTC-1              7           4              2                   1         4
#>                    pseudobulk_UMAP markers_UMAP SpatialDE_PCA_spatial SpatialDE_pool_PCA_spatial
#>                          <integer>    <integer>             <integer>                  <integer>
#> AAACAACGAATAGTTC-1               3            1                     3                          5
#> AAACAAGTATCTCCCA-1               2            1                     4                          1
#> AAACAATCTACTAGCA-1               3            6                     3                          4
#> AAACACCAATAACTGC-1               7            1                     7                          6
#> AAACAGCTTTCAGAAG-1               1            1                     1                          2
#> ...                            ...          ...                   ...                        ...
#> TTGTTGTGTGTCAAGA-1               2            3                     1                          1
#> TTGTTTCACATCCAGG-1               6            1                     5                          5
#> TTGTTTCATTAGTCTA-1               4            2                     5                          5
#> TTGTTTCCATACAACT-1               3            1                     4                          8
#> TTGTTTGTGTAAATTC-1               2            6                     2                          2
#>                    HVG_PCA_spatial pseudobulk_PCA_spatial markers_PCA_spatial
#>                          <integer>              <integer>           <integer>
#> AAACAACGAATAGTTC-1               5                      4                   3
#> AAACAAGTATCTCCCA-1               2                      3                   1
#> AAACAATCTACTAGCA-1               3                      5                   7
#> AAACACCAATAACTGC-1               7                      7                   7
#> AAACAGCTTTCAGAAG-1               1                      1                   2
#> ...                            ...                    ...                 ...
#> TTGTTGTGTGTCAAGA-1               4                      2                   6
#> TTGTTTCACATCCAGG-1               5                      4                   4
#> TTGTTTCATTAGTCTA-1               5                      4                   2
#> TTGTTTCCATACAACT-1               2                      2                   4
#> TTGTTTGTGTAAATTC-1               1                      3                   2
#>                    SpatialDE_UMAP_spatial SpatialDE_pool_UMAP_spatial HVG_UMAP_spatial
#>                                 <integer>                   <integer>        <integer>
#> AAACAACGAATAGTTC-1                      1                           1                1
#> AAACAAGTATCTCCCA-1                      2                           2                1
#> AAACAATCTACTAGCA-1                      2                           1                4
#> AAACACCAATAACTGC-1                      7                           8                7
#> AAACAGCTTTCAGAAG-1                      8                           2                8
#> ...                                   ...                         ...              ...
#> TTGTTGTGTGTCAAGA-1                      6                           1                4
#> TTGTTTCACATCCAGG-1                      5                           5                5
#> TTGTTTCATTAGTCTA-1                      5                           5                8
#> TTGTTTCCATACAACT-1                      2                           3                2
#> TTGTTTGTGTAAATTC-1                      1                           2                1
#>                    pseudobulk_UMAP_spatial markers_UMAP_spatial spatialLIBD ManualAnnotation
#>                                  <integer>            <integer>    <factor>      <character>
#> AAACAACGAATAGTTC-1                       1                    1          L1               NA
#> AAACAAGTATCTCCCA-1                       2                    1          L3               NA
#> AAACAATCTACTAGCA-1                       1                    6          L1               NA
#> AAACACCAATAACTGC-1                       8                    1          WM               NA
#> AAACAGCTTTCAGAAG-1                       1                    1          L6               NA
#> ...                                    ...                  ...         ...              ...
#> TTGTTGTGTGTCAAGA-1                       1                    1          L6               NA
#> TTGTTTCACATCCAGG-1                       8                    1          WM               NA
#> TTGTTTCATTAGTCTA-1                       1                    1          WM               NA
#> TTGTTTCCATACAACT-1                       4                    7          L6               NA
#> TTGTTTGTGTAAATTC-1                       2                    5          L1               NA
#>                    in_tissue array_row array_col imported_GraphBased imported_spatialLIBD
#>                    <logical> <integer> <integer>           <integer>          <character>
#> AAACAACGAATAGTTC-1      TRUE         0        16                   6                   L1
#> AAACAAGTATCTCCCA-1      TRUE        50       102                   3                   L3
#> AAACAATCTACTAGCA-1      TRUE         3        43                   2                   L1
#> AAACACCAATAACTGC-1      TRUE        59        19                   5                   WM
#> AAACAGCTTTCAGAAG-1      TRUE        43         9                   1                   L6
#> ...                      ...       ...       ...                 ...                  ...
#> TTGTTGTGTGTCAAGA-1      TRUE        31        77                   5                   L6
#> TTGTTTCACATCCAGG-1      TRUE        58        42                   2                   WM
#> TTGTTTCATTAGTCTA-1      TRUE        60        30                   2                   WM
#> TTGTTTCCATACAACT-1      TRUE        45        27                   7                   L6
#> TTGTTTGTGTAAATTC-1      TRUE         7        51                   1                   L1