Update the image for one sample

Edit the image with img_edit() then update the imgData().

img_update(
  spe,
  sampleid,
  image_id = "lowres",
  new_image_id = paste0("edited_", image_id),
  overwrite = FALSE,
  ...
)

Arguments

spe

A SpatialExperiment-class object. See fetch_data() for how to download some example objects or read10xVisiumWrapper() to read in spaceranger --count output files and build your own spe object.

sampleid

A character(1) specifying which sample to plot from colData(spe)$sample_id (formerly colData(spe)$sample_name).

image_id

A character(1) with the name of the image ID you want to use in the background.

new_image_id

A character(1) specifying the new image_id to use.

overwrite

A logical(1) specifying whether to overwrite the image_id if it already exists.

...

Parameters passed to img_edit().

Value

A SpatialExperiment-class object with an updated imgData() slot.

See also

Examples

if (enough_ram()) {
    ## Obtain the necessary data
    if (!exists("spe")) spe <- fetch_data("spe")

    ## Reduce brightness to 25% and update the imgData()
    imgData(img_update(spe, sampleid = "151507", brightness = 25))
}
#> 2024-07-19 16:09:51.600009 loading file /github/home/.cache/R/BiocFileCache/d682f29a8c2_Human_DLPFC_Visium_processedData_sce_scran_spatialLIBD.Rdata%3Fdl%3D1
#> DataFrame with 13 rows and 4 columns
#>       sample_id      image_id   data scaleFactor
#>     <character>   <character> <list>   <numeric>
#> 1        151507        lowres   ####   0.0450045
#> 2        151508        lowres   ####   0.0450045
#> 3        151509        lowres   ####   0.0450045
#> 4        151510        lowres   ####   0.0450045
#> 5        151669        lowres   ####   0.0450045
#> ...         ...           ...    ...         ...
#> 9        151673        lowres   ####   0.0450045
#> 10       151674        lowres   ####   0.0450045
#> 11       151675        lowres   ####   0.0450045
#> 12       151676        lowres   ####   0.0450045
#> 13       151507 edited_lowres   ####   0.0450045