Spatial registration: compute ANOVA statistics

This function computes the gene ANOVA F-statistics (at least one group is different from the rest). These F-statistics can be used for spatial registration with layer_stat_cor() and related functions. Although, they are more typically used for identifying ANOVA-marker genes.

registration_stats_anova(
  sce_pseudo,
  block_cor,
  covars = NULL,
  var_registration = "registration_variable",
  var_sample_id = "registration_sample_id",
  gene_ensembl = NULL,
  gene_name = NULL,
  prefix = ""
)

Arguments

sce_pseudo

The output of registration_pseudobulk().

block_cor

A numeric(1) computed with registration_block_cor().

covars

A character() with names of sample-level covariates.

var_registration

A character(1) specifying the colData(sce_pseudo) variable of interest against which will be used for computing the relevant statistics.

var_sample_id

A character(1) specifying the colData(sce_pseudo) variable with the sample ID.

gene_ensembl

A character(1) specifying the rowData(sce_pseudo) column with the ENSEMBL gene IDs. This will be used by layer_stat_cor().

gene_name

A character(1) specifying the rowData(sce_pseudo) column with the gene names (symbols).

prefix

A character(1) specifying the prefix to use for the F-statistics column. This is particularly useful if you will run this function more than once and want to be able to merge the results.

Value

A data.frame() with the ANOVA statistical results. This is similar to fetch_data("modeling_results")$anova.

See also

Examples

example("registration_block_cor", package = "spatialLIBD")
#> 
#> rgst__master > example("registration_model", package = "spatialLIBD")
#> 
#> rgstr_master > example("registration_pseudobulk", package = "spatialLIBD")
#> 
#> rgstr_master > ## Ensure reproducibility of example data
#> rgstr_master > set.seed(20220907)
#> 
#> rgstr_master > ## Generate example data
#> rgstr_master > sce <- scuttle::mockSCE()
#> 
#> rgstr_master > ## Add some sample IDs
#> rgstr_master > sce$sample_id <- sample(LETTERS[1:5], ncol(sce), replace = TRUE)
#> 
#> rgstr_master > ## Add a sample-level covariate: age
#> rgstr_master > ages <- rnorm(5, mean = 20, sd = 4)
#> 
#> rgstr_master > names(ages) <- LETTERS[1:5]
#> 
#> rgstr_master > sce$age <- ages[sce$sample_id]
#> 
#> rgstr_master > ## Add gene-level information
#> rgstr_master > rowData(sce)$ensembl <- paste0("ENSG", seq_len(nrow(sce)))
#> 
#> rgstr_master > rowData(sce)$gene_name <- paste0("gene", seq_len(nrow(sce)))
#> 
#> rgstr_master > ## Pseudo-bulk
#> rgstr_master > sce_pseudo <- registration_pseudobulk(sce, "Treatment", "sample_id", c("age"))
#> 2022-09-08 12:34:25 make pseudobulk object
#> Warning: 'use.altexps=' is deprecated.
#> Use 'applySCE(x, aggregateAcrossCells)' instead.
#> 2022-09-08 12:34:25 drop lowly expressed genes
#> 2022-09-08 12:34:25 normalize expression
#> 
#> rgstr_master > colData(sce_pseudo)
#> DataFrame with 10 rows and 8 columns
#>          Mutation_Status  Cell_Cycle   Treatment   sample_id       age
#>              <character> <character> <character> <character> <numeric>
#> A_treat1              NA          NA      treat1           A   19.1872
#> B_treat1              NA          NA      treat1           B   25.3496
#> C_treat1              NA          NA      treat1           C   24.1802
#> D_treat1              NA          NA      treat1           D   15.5211
#> E_treat1              NA          NA      treat1           E   20.9701
#> A_treat2              NA          NA      treat2           A   19.1872
#> B_treat2              NA          NA      treat2           B   25.3496
#> C_treat2              NA          NA      treat2           C   24.1802
#> D_treat2              NA          NA      treat2           D   15.5211
#> E_treat2              NA          NA      treat2           E   20.9701
#>          registration_variable registration_sample_id    ncells
#>                    <character>            <character> <integer>
#> A_treat1                treat1                      A        18
#> B_treat1                treat1                      B        19
#> C_treat1                treat1                      C        18
#> D_treat1                treat1                      D        24
#> E_treat1                treat1                      E        22
#> A_treat2                treat2                      A        18
#> B_treat2                treat2                      B        23
#> C_treat2                treat2                      C        20
#> D_treat2                treat2                      D        20
#> E_treat2                treat2                      E        18
#> 
#> rgstr_master > registration_mod <- registration_model(sce_pseudo, "age")
#> 2022-09-08 12:34:25 create model matrix
#> 
#> rgstr_master > head(registration_mod)
#>          registration_variabletreat1 registration_variabletreat2      age
#> A_treat1                           1                           0 19.18719
#> B_treat1                           1                           0 25.34965
#> C_treat1                           1                           0 24.18019
#> D_treat1                           1                           0 15.52107
#> E_treat1                           1                           0 20.97006
#> A_treat2                           0                           1 19.18719
#> 
#> rgst__master > block_cor <- registration_block_cor(sce_pseudo, registration_mod)
#> 2022-09-08 12:34:25 run duplicateCorrelation()
results_anova <- registration_stats_anova(sce_pseudo,
    block_cor, "age",
    gene_ensembl = "ensembl", gene_name = "gene_name", prefix = "example"
)
#> 2022-09-08 12:34:25 computing F-statistics
head(results_anova)
#>   example_f_stat example_p_value example_fdr example_AveExpr ensembl  gene
#> 1      0.1814036      0.83707093   0.9894421        5.603911   ENSG1 gene1
#> 2      0.2308187      0.79843629   0.9894421        3.965248   ENSG2 gene2
#> 3      0.6073113      0.56568413   0.9894421       10.744697   ENSG3 gene3
#> 4      0.6422112      0.54860918   0.9894421        5.794754   ENSG4 gene4
#> 5      7.3540538      0.01277300   0.8562882        3.336278   ENSG5 gene5
#> 6      4.2767610      0.04943749   0.9654248        5.431549   ENSG6 gene6

## Note that you can merge multiple of these data.frames if you run this
## function for different sets. For example, maybe you drop one group
## before pseudo-bulking if you know that there are many differences between
## that group and others. For example, we have dropped the white matter (WM)
## prior to computing ANOVA F-statistics.