Spatial registration: compute enrichment statistics

This function computes the gene enrichment t-statistics (one group > the rest). These t-statistics are the ones typically used for spatial registration with layer_stat_cor() and related functions.

registration_stats_enrichment(
  sce_pseudo,
  block_cor,
  covars = NULL,
  var_registration = "registration_variable",
  var_sample_id = "registration_sample_id",
  gene_ensembl = NULL,
  gene_name = NULL
)

Arguments

sce_pseudo

The output of registration_pseudobulk().

block_cor

A numeric(1) computed with registration_block_cor().

covars

A character() with names of sample-level covariates.

var_registration

A character(1) specifying the colData(sce_pseudo) variable of interest against which will be used for computing the relevant statistics.

var_sample_id

A character(1) specifying the colData(sce_pseudo) variable with the sample ID.

gene_ensembl

A character(1) specifying the rowData(sce_pseudo) column with the ENSEMBL gene IDs. This will be used by layer_stat_cor().

gene_name

A character(1) specifying the rowData(sce_pseudo) column with the gene names (symbols).

Value

A data.frame() with the enrichment statistical results. This is similar to fetch_data("modeling_results")$enrichment.

See also

Examples

example("registration_block_cor", package = "spatialLIBD")
#> 
#> rgst__master > block_cor <- registration_block_cor(sce_pseudo, mod)
#> 2022-09-08 10:13:47 run duplicateCorrelation()
results_enrichment <- registration_stats_enrichment(sce_pseudo,
    block_cor, "age",
    gene_ensembl = "ensembl", gene_name = "gene_name"
)
#> 2022-09-08 10:13:48 run enrichment statistics
#> 2022-09-08 10:13:48 extract and reformat enrichment results
head(results_enrichment)
#>           t_stat_treat1 t_stat_treat2 p_value_treat1 p_value_treat2 fdr_treat1
#> Gene_0001    -0.9733723     0.9733723     0.35568479     0.35568479  0.9975299
#> Gene_0002     0.4150539    -0.4150539     0.68777696     0.68777696  0.9975299
#> Gene_0003    -0.4226473     0.4226473     0.68242530     0.68242530  0.9975299
#> Gene_0004     2.2869931    -2.2869931     0.04787877     0.04787877  0.9975299
#> Gene_0005     0.6456216    -0.6456216     0.53456193     0.53456193  0.9975299
#> Gene_0006    -0.4108131     0.4108131     0.69077385     0.69077385  0.9975299
#>           fdr_treat2 ensembl  gene
#> Gene_0001  0.9975299   ENSG1 gene1
#> Gene_0002  0.9975299   ENSG2 gene2
#> Gene_0003  0.9975299   ENSG3 gene3
#> Gene_0004  0.9975299   ENSG4 gene4
#> Gene_0005  0.9975299   ENSG5 gene5
#> Gene_0006  0.9975299   ENSG6 gene6