Spatial registration: compute pairwise statistics

This function computes the gene pairwise t-statistics (one group > another, for all combinations). These t-statistics can be used for spatial registration with layer_stat_cor() and related functions. Although, they are more typically used for identifying pairwise-marker genes.

registration_stats_pairwise(
  sce_pseudo,
  registration_model,
  block_cor,
  var_sample_id = "registration_sample_id",
  gene_ensembl = NULL,
  gene_name = NULL
)

Arguments

sce_pseudo

The output of registration_pseudobulk().

registration_model

The output from registration_model().

block_cor

A numeric(1) computed with registration_block_cor().

var_sample_id

A character(1) specifying the colData(sce_pseudo) variable with the sample ID.

gene_ensembl

A character(1) specifying the rowData(sce_pseudo) column with the ENSEMBL gene IDs. This will be used by layer_stat_cor().

gene_name

A character(1) specifying the rowData(sce_pseudo) column with the gene names (symbols).

Value

A data.frame() with the pairwise statistical results. This is similar to fetch_data("modeling_results")$pairwise.

See also

Examples

example("registration_block_cor", package = "spatialLIBD")
#> Warning: no help found for ‘registration_block_cor’
results_pairwise <- registration_stats_pairwise(sce_pseudo,
    registration_mod, block_cor,
    gene_ensembl = "ensembl", gene_name = "gene_name"
)
#> Error in registration_stats_pairwise(sce_pseudo, registration_mod, block_cor,     gene_ensembl = "ensembl", gene_name = "gene_name"): could not find function "registration_stats_pairwise"
head(results_pairwise)
#> Error in h(simpleError(msg, call)): error in evaluating the argument 'x' in selecting a method for function 'head': object 'results_pairwise' not found