Sample spatial cluster visualization workhorse function

This function visualizes the clusters for one given sample at the spot-level using (by default) the histology information on the background. This is the function that does all the plotting behind sce_image_clus(). To visualize gene-level (or any continuous variable) use sce_image_gene_p().

sce_image_clus_p(sce, d, clustervar, sampleid, colors, spatial, title)

Arguments

sce
Defaults to the output of fetch_data(type = 'sce'). This is a SingleCellExperiment object with the spot-level Visium data and information required for visualizing the histology. See fetch_data() for more details.

d
A data.frame with the sample-level information. This is typically obtained using as.data.frame(colData(sce)).

clustervar
A character(1) with the name of the colData(sce) column that has the cluster values.

sampleid
A character(1) specifying which sample to plot from colData(sce)$sample_name.

colors
A vector of colors to use for visualizing the clusters from clustervar. If the vector has names, then those should match the values of clustervar.

spatial
A logical(1) indicating whether to include the histology layer from geom_spatial(). If you plan to use ggplotly() then it's best to set this to FALSE.

title
The title for the plot.

Value

A ggplot2 object.

See also

Examples

if (enough_ram()) {
    ## Obtain the necessary data
    if (!exists("sce")) sce <- fetch_data("sce")
    sce_sub <- sce[, sce$sample_name == "151673"]

    ## Use the manual color palette by Lukas M Weber
    ## Don't plot the histology information
    sce_image_clus_p(
        sce = sce_sub,
        d = as.data.frame(colData(sce_sub)),
        clustervar = "layer_guess_reordered",
        sampleid = "151673",
        colors = libd_layer_colors,
        title = "151673 LIBD Layers",
        spatial = FALSE
    )

    ## Clean up
    rm(sce_sub)
}

#> snapshotDate(): 2020-09-24

#> 2020-09-28 18:58:25 loading file /tmp/RtmpE2FbEl/BiocFileCache/6a2b41bc535e_Human_DLPFC_Visium_processedData_sce_scran_spatialLIBD.Rdata%3Fdl%3D1