Sample spatial gene visualization workhorse function

This function visualizes the gene expression stored in assays(sce) or any continuous variable stored in colData(sce) for one given sample at the spot-level using (by default) the histology information on the background. This is the function that does all the plotting behind sce_image_gene(). To visualize clusters (or any discrete variable) use sce_image_clus_p().

sce_image_gene_p(sce, d, sampleid, spatial, title, viridis = TRUE)

Arguments

sce
Defaults to the output of fetch_data(type = 'sce'). This is a SingleCellExperiment object with the spot-level Visium data and information required for visualizing the histology. See fetch_data() for more details.

d
A data.frame with the sample-level information. This is typically obtained using as.data.frame(colData(sce)). The data.frame has to contain a column with the continuous variable data to plot stored under d$COUNT.

sampleid
A character(1) specifying which sample to plot from colData(sce)$sample_name.

spatial
A logical(1) indicating whether to include the histology layer from geom_spatial(). If you plan to use ggplotly() then it's best to set this to FALSE.

title
The title for the plot.

viridis
A logical(1) whether to use the color-blind friendly palette from viridis or the color palette used in the paper that was chosen for contrast when visualizing the data on top of the histology image. One issue is being able to differentiate low values from NA ones due to the purple-ish histology information that is dependent on cell density.

Value

A ggplot2 object.

See also

Examples

if (enough_ram()) {
    ## Obtain the necessary data
    if (!exists("sce")) sce <- fetch_data("sce")

    ## Prepare the data for the plotting function
    sce_sub <- sce[, sce$sample_name == "151673"]
    df <- as.data.frame(colData(sce_sub))
    df$COUNT <- df$expr_chrM_ratio

    ## Use the manual color palette by Lukas M Weber
    ## Don't plot the histology information
    sce_image_gene_p(
        sce = sce_sub,
        d = df,
        sampleid = "151673",
        title = "151673 chrM expr ratio",
        spatial = FALSE
    )
    

    ## Or you can do this with a VisiumEsperiment object
    ve_sub <- sce_to_ve(sce_sub)
    df2 <- colData(ve_sub)
    df2$COUNT <- df2$expr_chrM_ratio
    
    sce_image_gene_p(
        sce = ve_sub,
        d = df2,
        sampleid = "151673",
        title = "151673 chrM expr ratio",
        spatial = FALSE
    )
    

    ## Clean up
    rm(sce_sub)
    rm(ve_sub)
}

#> snapshotDate(): 2020-10-02

#> 2020-10-26 22:57:48 loading file /tmp/RtmpdWbfFC/BiocFileCache/c292e303f00_Human_DLPFC_Visium_processedData_sce_scran_spatialLIBD.Rdata%3Fdl%3D1