Extract significant genes

From the layer-level modeling results, this function extracts the top n significant genes. This is the workhorse function used by sig_genes_extract_all() through which we obtain the information that can then be used by functions such as layer_boxplot() for constructing informative titles.

sig_genes_extract(
  n = 10,
  modeling_results = fetch_data(type = "modeling_results"),
  model_type = names(modeling_results)[1],
  reverse = FALSE,
  sce_layer = fetch_data(type = "sce_layer")
)

Arguments

n
The number of the top ranked genes to extract.

modeling_results
Defaults to the output of fetch_data(type = 'modeling_results'). This is a list of tables with the columns f_stat_* or t_stat_* as well as p_value_* and fdr_* plus ensembl. The column name is used to extract the statistic results, the p-values, and the FDR adjusted p-values. Then the ensembl column is used for matching in some cases. See fetch_data() for more details.

model_type
A named element of the modeling_results list. By default that is either enrichment for the model that tests one human brain layer against the rest (one group vs the rest), pairwise which compares two layers (groups) denoted by layerA-layerB such that layerA is greater than layerB, and anova which determines if any layer (group) is different from the rest adjusting for the mean expression level. The statistics for enrichment and pairwise are t-statistics while the anova model ones are F-statistics.

reverse
A logical(1) indicating whether to multiply by -1 the input statistics and reverse the layerA-layerB column names (using the -) into layerB-layerA.

sce_layer
Defaults to the output of fetch_data(type = 'sce_layer'). This is a SingleCellExperiment object with the spot-level Visium data compressed via pseudo-bulking to the layer-level (group-level) resolution. See fetch_data() for more details.

Value

A data.frame() with the top n significant genes (as ordered by their statistics in decreasing order) in long format. The specific columns are described further in the vignette.

References

Adapted from https://github.com/LieberInstitute/HumanPilot/blob/master/Analysis/Layer_Guesses/layer_specificity_functions.R

See also

Examples

## Obtain the necessary data
if (!exists("modeling_results")) {
      modeling_results <- fetch_data(type = "modeling_results")
  }

#> snapshotDate(): 2020-04-03

#> Loading objects:
#>   modeling_results

if (!exists("sce_layer")) sce_layer <- fetch_data(type = "sce_layer")

#> snapshotDate(): 2020-04-03

#> Loading objects:
#>   sce_layer

## anova top 10 genes
sig_genes_extract(
    modeling_results = modeling_results,
    sce_layer = sce_layer
)

#>    top model_type test     gene     stat         pval          fdr gene_index
#> 1    1      anova full    KRT17 142.6223 1.047345e-37 2.338827e-33      18321
#> 2    2      anova full SERPINE2 130.0159 2.154786e-36 2.405926e-32       3621
#> 3    3      anova full     GFAP 127.3117 4.267922e-36 3.176899e-32      18436
#> 4    4      anova full    CBLN4 118.0481 4.923904e-35 2.748892e-31      19844
#> 5    5      anova full   VSTM2A 111.0341 3.530274e-34 1.576691e-30       8315
#> 6    6      anova full  CAMK2N1 108.9155 6.542909e-34 2.435162e-30        308
#> 7    7      anova full   HS3ST4 104.9437 2.142606e-33 6.049917e-30      16991
#> 8    8      anova full     MYRF 104.9059 2.167361e-33 6.049917e-30      11871
#> 9    9      anova full    BCAS1 103.3046 3.537055e-33 8.776220e-30      19839
#> 10  10      anova full     MOBP 100.9366 7.389973e-33 1.650255e-29       3978
#> 11   1      anova noWM SERPINE2 209.6240 9.525498e-37 2.127139e-32       3621
#> 12   2      anova noWM  B3GALT2 181.2393 5.843659e-35 4.749316e-31       1809
#> 13   3      anova noWM    KRT17 180.6747 6.380345e-35 4.749316e-31      18321
#> 14   4      anova noWM    EFHD2 163.2460 1.099248e-33 5.247556e-30        232
#> 15   5      anova noWM  CAMK2N1 162.8576 1.174949e-33 5.247556e-30        308
#> 16   6      anova noWM     HOPX 157.1804 3.164231e-33 1.177674e-29       5291
#> 17   7      anova noWM   CLSTN2 148.4278 1.558137e-32 4.970680e-29       4638
#> 18   8      anova noWM   TUBA1B 135.6197 1.890890e-31 4.918869e-28      13893
#> 19   9      anova noWM   HS3ST2 135.3872 1.982438e-31 4.918869e-28      16964
#> 20  10      anova noWM     ETV1 130.0174 6.033934e-31 1.347438e-27       8095
#>            ensembl
#> 1  ENSG00000128422
#> 2  ENSG00000135919
#> 3  ENSG00000131095
#> 4  ENSG00000054803
#> 5  ENSG00000170419
#> 6  ENSG00000162545
#> 7  ENSG00000182601
#> 8  ENSG00000124920
#> 9  ENSG00000064787
#> 10 ENSG00000168314
#> 11 ENSG00000135919
#> 12 ENSG00000162630
#> 13 ENSG00000128422
#> 14 ENSG00000142634
#> 15 ENSG00000162545
#> 16 ENSG00000171476
#> 17 ENSG00000158258
#> 18 ENSG00000123416
#> 19 ENSG00000122254
#> 20 ENSG00000006468

## Extract all genes
sig_genes_extract(
    modeling_results = modeling_results,
    sce_layer = sce_layer,
    n = nrow(sce_layer)
)

#>     top model_type test     gene      stat         pval          fdr gene_index
#> 1     1      anova full    KRT17 142.62226 1.047345e-37 2.338827e-33      18321
#> 2     2      anova full SERPINE2 130.01590 2.154786e-36 2.405926e-32       3621
#> 3     3      anova full     GFAP 127.31169 4.267922e-36 3.176899e-32      18436
#> 4     4      anova full    CBLN4 118.04809 4.923904e-35 2.748892e-31      19844
#> 5     5      anova full   VSTM2A 111.03407 3.530274e-34 1.576691e-30       8315
#> 6     6      anova full  CAMK2N1 108.91547 6.542909e-34 2.435162e-30        308
#> 7     7      anova full   HS3ST4 104.94373 2.142606e-33 6.049917e-30      16991
#> 8     8      anova full     MYRF 104.90591 2.167361e-33 6.049917e-30      11871
#> 9     9      anova full    BCAS1 103.30459 3.537055e-33 8.776220e-30      19839
#> 10   10      anova full     MOBP 100.93656 7.389973e-33 1.650255e-29       3978
#> 11   11      anova full    EFHD2  95.11926 4.830189e-32 9.805724e-29        232
#> 12   12      anova full      MBP  93.31013 8.841183e-32 1.645271e-28      19352
#> 13   13      anova full     PLP1  90.81919 2.067407e-31 3.551327e-28       9433
#> 14   14      anova full     TLE4  89.57378 3.185778e-31 5.081544e-28      10842
#> 15   15      anova full    SCN1B  89.29950 3.506551e-31 5.220320e-28      20704
#> 16   16      anova full   HS3ST2  88.80268 4.174799e-31 5.826714e-28      16964
#> 17   17      anova full   HPCAL1  87.92038 5.702797e-31 7.491127e-28       2313
#> 18   18      anova full   FRMPD2  87.32346 7.053483e-31 8.394075e-28      12882
#> 19   19      anova full     HOPX  87.28856 7.141974e-31 8.394075e-28       5291
#> 20   20      anova full     CTGF  86.88731 8.245258e-31 9.206242e-28       7750
#> 21   21      anova full   CARTPT  86.45410 9.634771e-31 1.024543e-27       6121
#> 22   22      anova full      MAG  85.73052 1.251635e-30 1.242947e-27      20716
#> 23   23      anova full     MT1G  85.49982 1.361082e-30 1.242947e-27      17246
#> 24   24      anova full    LAMP5  85.49858 1.361696e-30 1.242947e-27      19480
#> 25   25      anova full    PARM1  85.43907 1.391504e-30 1.242947e-27       5344
#> 26   26      anova full     CNR1  82.78401 3.706832e-30 3.183741e-27       7564
#> 27   27      anova full     SNCG  81.49962 6.013272e-30 4.973422e-27      13092
#> 28   28      anova full   CLDND1  79.53535 1.276556e-29 1.018099e-26       4350
#> 29   29      anova full   OPALIN  79.24022 1.431399e-29 1.102226e-26      13171
#> 30   30      anova full   NECAB2  78.87941 1.647264e-29 1.226169e-26      17538
#> 31   31      anova full    RGS12  78.67761 1.782330e-29 1.283910e-26       5059
#> 32   32      anova full     RORB  77.71670 2.600133e-29 1.814486e-26      10825
#> 33   33      anova full     CRYM  75.38463 6.612801e-29 4.474863e-26      16941
#> 34   34      anova full       TF  74.48048 9.559810e-29 6.278827e-26       4598
#> 35   35      anova full    SEPT4  74.28070 1.037625e-28 6.620342e-26      18627
#> 36   36      anova full      MAL  73.88124 1.223066e-28 7.392804e-26       2851
#> 37   37      anova full   TUBA1B  73.87759 1.224906e-28 7.392804e-26      13893
#> 38   38      anova full    LPAR1  73.75946 1.286128e-28 7.558034e-26      11021
#> 39   39      anova full     ERMN  73.66065 1.339754e-28 7.671295e-26       3181
#> 40   40      anova full    EVI2A  72.84319 1.881777e-28 1.025818e-25      18141
#> 41   41      anova full  B3GALT2  72.84110 1.883415e-28 1.025818e-25       1809
#> 42   42      anova full    ERBB3  72.33845 2.324733e-28 1.236038e-25      14064
#> 43   43      anova full    PVALB  72.26690 2.395696e-28 1.244146e-25      21800
#> 44   44      anova full   SOWAHA  70.87215 4.327060e-28 2.196081e-25       6395
#> 45   45      anova full PPP1R14A  70.24044 5.674096e-28 2.765105e-25      20817
#> 46   46      anova full    OLFM1  70.23154 5.695886e-28 2.765105e-25      11296
#> 47   47      anova full   CLSTN2  70.07161 6.102471e-28 2.899453e-25       4638
#> 48   48      anova full      SHB  69.98933 6.323153e-28 2.941715e-25      10755
#> 49   49      anova full    SMYD2  69.81673 6.813144e-28 3.104986e-25       1994
#> 50   50      anova full   CLDN11  69.69652 7.177327e-28 3.205538e-25       4790
#> 51   51      anova full    CNDP1  69.23235 8.782105e-28 3.845356e-25      19335
#> 52   52      anova full     SNCB  68.96081 9.887700e-28 4.246197e-25       6771
#> 53   53      anova full   NAP1L5  68.85738 1.034557e-27 4.358997e-25       5434
#> 54   54      anova full    TPPP3  68.53788 1.190263e-27 4.922180e-25      17341
#> 55   55      anova full    PDZD2  68.33889 1.299236e-27 5.275134e-25       5932
#> 56   56      anova full      MOG  68.06957 1.463276e-27 5.835073e-25       7082
#> 57   57      anova full    S1PR5  67.98191 1.521144e-27 5.959415e-25      20298
#> 58   58      anova full      CNP  67.60795 1.795712e-27 6.913801e-25      18334
#> 59   59      anova full    STMN1  67.55910 1.835160e-27 6.945925e-25        390
#> 60   60      anova full   PCDH17  67.41889 1.953426e-27 7.270326e-25      14929
#> 61   61      anova full     JAM3  67.15544 2.197292e-27 8.043888e-25      12600
#> 62   62      anova full     SPP1  67.11037 2.242055e-27 8.075377e-25       5424
#> 63   63      anova full     NSG2  66.84392 2.526499e-27 8.836315e-25       6749
#> 64   64      anova full   RASSF2  66.83867 2.532462e-27 8.836315e-25      19455
#> 65   65      anova full     AQP1  66.58571 2.837625e-27 9.748771e-25       8201
#> 66   66      anova full     KLK6  66.37700 3.117757e-27 1.054888e-24      21228
#> 67   67      anova full    PCDH8  66.28215 3.254313e-27 1.084658e-24      14927
#> 68   68      anova full      CD9  66.16132 3.437214e-27 1.128771e-24      13536
#> 69   69      anova full    HSPA2  66.01644 3.670548e-27 1.187928e-24      15448
#> 70   70      anova full    LAMP2  65.92895 3.819274e-27 1.208047e-24       9511
#> 71   71      anova full  ADAMTS4  65.91652 3.840907e-27 1.208047e-24       1600
#> 72   72      anova full   NKX6-2  65.69436 4.249409e-27 1.317966e-24      13454
#> 73   73      anova full    RXFP1  65.65748 4.321441e-27 1.321946e-24       5706
#> 74   74      anova full    NDRG1  65.45209 4.746138e-27 1.432243e-24      10404
#> 75   75      anova full     CUX2  65.04134 5.728994e-27 1.692984e-24      14444
#> 76   76      anova full   HS6ST3  65.02892 5.761802e-27 1.692984e-24      14990
#> 77   77      anova full    CRYAB  64.96607 5.930678e-27 1.719974e-24      12390
#> 78   78      anova full    EPHA4  64.72078 6.640039e-27 1.901009e-24       3606
#> 79   79      anova full   CAMK2D  64.59040 7.052032e-27 1.993404e-24       5548
#> 80   80      anova full     MT2A  64.50088 7.350042e-27 2.051672e-24      17241
#> 81   81      anova full    SATB1  64.19407 8.473373e-27 2.336036e-24       3886
#> 82   82      anova full    CALB2  63.95806 9.456666e-27 2.551130e-24      17428
#> 83   83      anova full     NEFL  63.95231 9.482055e-27 2.551130e-24       9854
#> 84   84      anova full  TMEM144  63.91723 9.638339e-27 2.562306e-24       5705
#> 85   85      anova full  SELENOP  63.78713 1.024149e-26 2.669139e-24       5993
#> 86   86      anova full    SNRPN  63.77925 1.027925e-26 2.669139e-24      15855
#> 87   87      anova full   RNASE1  63.49234 1.175599e-26 3.017506e-24      15115
#> 88   88      anova full    CLIC4  63.10849 1.407984e-26 3.572920e-24        372
#> 89   89      anova full     NFIA  63.03233 1.459447e-26 3.661900e-24        845
#> 90   90      anova full     NSG1  62.72375 1.688565e-26 4.189705e-24       5069
#> 91   91      anova full    CRLF1  62.52176 1.858283e-26 4.516390e-24      20551
#> 92   92      anova full    FABP7  62.51905 1.860677e-26 4.516390e-24       7713
#> 93   93      anova full     RCN1  62.03260 2.345867e-26 5.632854e-24      11682
#> 94   94      anova full    VAMP1  61.84644 2.564412e-26 6.092115e-24      13544
#> 95   95      anova full     HHIP  61.80629 2.614236e-26 6.121400e-24       5646
#> 96   96      anova full      CA4  61.79251 2.631563e-26 6.121400e-24      18652
#> 97   97      anova full    NR4A2  61.51882 3.001264e-26 6.909405e-24       3178
#> 98   98      anova full   CARNS1  61.43900 3.118843e-26 7.106825e-24      12099
#> 99   99      anova full      SCD  61.37812 3.211695e-26 7.244481e-24      13221
#> 100 100      anova full    PMP22  60.92389 4.000660e-26 8.933874e-24      17940
#> 101 101      anova full  NEUROD6  60.22879 5.613816e-26 1.241209e-23       8204
#> 102 102      anova full  TMEM125  60.20007 5.693345e-26 1.246452e-23        660
#> 103 103      anova full     ASS1  59.82893 6.831717e-26 1.481156e-23      11234
#> 104 104      anova full   FAM84A  59.68038 7.350694e-26 1.578349e-23       2335
#> 105 105      anova full   ELOVL1  59.52045 7.954890e-26 1.691816e-23        666
#> 106 106      anova full  RETREG1  59.44490 8.257848e-26 1.739679e-23       5910
#> 107 107      anova full    MEF2C  59.36216 8.603284e-26 1.795513e-23       6227
#> 108 108      anova full    CERS2  59.28061 8.958301e-26 1.852295e-23       1338
#> 109 109      anova full   PTP4A2  59.08478 9.873643e-26 2.022829e-23        487
#> 110 110      anova full     NEFM  58.86242 1.103046e-25 2.239283e-23       9853
#> 111 111      anova full     GJC2  58.49740 1.324042e-25 2.663710e-23       2083
#>             ensembl
#> 1   ENSG00000128422
#> 2   ENSG00000135919
#> 3   ENSG00000131095
#> 4   ENSG00000054803
#> 5   ENSG00000170419
#> 6   ENSG00000162545
#> 7   ENSG00000182601
#> 8   ENSG00000124920
#> 9   ENSG00000064787
#> 10  ENSG00000168314
#> 11  ENSG00000142634
#> 12  ENSG00000197971
#> 13  ENSG00000123560
#> 14  ENSG00000106829
#> 15  ENSG00000105711
#> 16  ENSG00000122254
#> 17  ENSG00000115756
#> 18  ENSG00000170324
#> 19  ENSG00000171476
#> 20  ENSG00000118523
#> 21  ENSG00000164326
#> 22  ENSG00000105695
#> 23  ENSG00000125144
#> 24  ENSG00000125869
#> 25  ENSG00000169116
#> 26  ENSG00000118432
#> 27  ENSG00000173267
#> 28  ENSG00000080822
#> 29  ENSG00000197430
#> 30  ENSG00000103154
#> 31  ENSG00000159788
#> 32  ENSG00000198963
#> 33  ENSG00000103316
#> 34  ENSG00000091513
#> 35  ENSG00000108387
#> 36  ENSG00000172005
#> 37  ENSG00000123416
#> 38  ENSG00000198121
#> 39  ENSG00000136541
#> 40  ENSG00000126860
#> 41  ENSG00000162630
#> 42  ENSG00000065361
#> 43  ENSG00000100362
#> 44  ENSG00000198944
#> 45  ENSG00000167641
#> 46  ENSG00000130558
#> 47  ENSG00000158258
#> 48  ENSG00000107338
#> 49  ENSG00000143499
#> 50  ENSG00000013297
#> 51  ENSG00000150656
#> 52  ENSG00000074317
#> 53  ENSG00000177432
#> 54  ENSG00000159713
#> 55  ENSG00000133401
#> 56  ENSG00000204655
#> 57  ENSG00000180739
#> 58  ENSG00000173786
#> 59  ENSG00000117632
#> 60  ENSG00000118946
#> 61  ENSG00000166086
#> 62  ENSG00000118785
#> 63  ENSG00000170091
#> 64  ENSG00000101265
#> 65  ENSG00000240583
#> 66  ENSG00000167755
#> 67  ENSG00000136099
#> 68  ENSG00000010278
#> 69  ENSG00000126803
#> 70  ENSG00000005893
#> 71  ENSG00000158859
#> 72  ENSG00000148826
#> 73  ENSG00000171509
#> 74  ENSG00000104419
#> 75  ENSG00000111249
#> 76  ENSG00000185352
#> 77  ENSG00000109846
#> 78  ENSG00000116106
#> 79  ENSG00000145349
#> 80  ENSG00000125148
#> 81  ENSG00000182568
#> 82  ENSG00000172137
#> 83  ENSG00000277586
#> 84  ENSG00000164124
#> 85  ENSG00000250722
#> 86  ENSG00000128739
#> 87  ENSG00000129538
#> 88  ENSG00000169504
#> 89  ENSG00000162599
#> 90  ENSG00000168824
#> 91  ENSG00000006016
#> 92  ENSG00000164434
#> 93  ENSG00000049449
#> 94  ENSG00000139190
#> 95  ENSG00000164161
#> 96  ENSG00000167434
#> 97  ENSG00000153234
#> 98  ENSG00000172508
#> 99  ENSG00000099194
#> 100 ENSG00000109099
#> 101 ENSG00000164600
#> 102 ENSG00000179178
#> 103 ENSG00000130707
#> 104 ENSG00000162981
#> 105 ENSG00000066322
#> 106 ENSG00000154153
#> 107 ENSG00000081189
#> 108 ENSG00000143418
#> 109 ENSG00000184007
#> 110 ENSG00000104722
#> 111 ENSG00000198835
#>  [ reached 'max' / getOption("max.print") -- omitted 44551 rows ]