Sample spatial gene visualization workhorse function

This function visualizes the gene expression stored in assays(spe) or any continuous variable stored in colData(spe) for one given sample at the spot-level using (by default) the histology information on the background. This is the function that does all the plotting behind vis_gene(). To visualize clusters (or any discrete variable) use vis_clus_p().

vis_gene_p(
  spe,
  d,
  sampleid,
  spatial,
  title,
  viridis = TRUE,
  image_id = "lowres",
  alpha = 1,
  cont_colors = if (viridis) viridisLite::viridis(21) else c("aquamarine4",
    "springgreen", "goldenrod", "red")
)

Arguments

spe
Defaults to the output of fetch_data(type = 'spe'). This is a SpatialExperiment-class object with the spot-level Visium data and information required for visualizing the histology. See fetch_data() for more details.

d
A data.frame with the sample-level information. This is typically obtained using colData(spe, spatialData = TRUE, spatialCoords = TRUE). The data.frame has to contain a column with the continuous variable data to plot stored under d$COUNT.

sampleid
A character(1) specifying which sample to plot from colData(spe)$sample_name.

spatial
A logical(1) indicating whether to include the histology layer from geom_spatial(). If you plan to use ggplotly() then it's best to set this to FALSE.

title
The title for the plot.

viridis
A logical(1) whether to use the color-blind friendly palette from viridis or the color palette used in the paper that was chosen for contrast when visualizing the data on top of the histology image. One issue is being able to differentiate low values from NA ones due to the purple-ish histology information that is dependent on cell density.

image_id
A character(1) with the name of the image ID you want to use in the background.

alpha
A numeric(1) in the [0, 1] range that specifies the transparency level of the data on the spots.

cont_colors
A character() vector of colors that supersedes the viridis argument.

Value

A ggplot2 object.

See also

Examples

if (enough_ram()) {
    ## Obtain the necessary data
    if (!exists("spe")) spe <- fetch_data("spe")

    ## Prepare the data for the plotting function
    spe_sub <- spe[, spe$sample_id == "151673"]
    df <- as.data.frame(colData(spe_sub, spatialData = TRUE, spatialCoords = TRUE), optional = TRUE)
    df$COUNT <- df$expr_chrM_ratio

    ## Use the manual color palette by Lukas M Weber
    ## Don't plot the histology information
    vis_gene_p(
        spe = spe_sub,
        d = df,
        sampleid = "151673",
        title = "151673 chrM expr ratio",
        spatial = FALSE
    )

    ## Clean up
    rm(spe_sub)
}
#> snapshotDate(): 2021-10-19
#> 2021-12-01 15:50:33 loading file /github/home/.cache/R/BiocFileCache/6d315fe0844_Human_DLPFC_Visium_processedData_sce_scran_spatialLIBD.Rdata%3Fdl%3D1