Evaluate the enrichment for a list of gene sets

Using the layer-level (group-level) data, this function evaluates whether list of gene sets (Ensembl gene IDs) are enrichment among the significant genes (FDR < 0.1 by default) genes for a given model type result.

gene_set_enrichment(
  gene_list,
  fdr_cut = 0.1,
  modeling_results = fetch_data(type = "modeling_results"),
  model_type = names(modeling_results)[1],
  reverse = FALSE
)

Arguments

gene_list
A named list object (could be a data.frame) where each element of the list is a character vector of Ensembl gene IDs.

fdr_cut
A numeric(1) specifying the FDR cutoff to use for determining significance among the

modeling_results
Defaults to the output of fetch_data(type = 'modeling_results'). This is a list of tables with the columns f_stat_* or t_stat_* as well as p_value_* and fdr_* plus ensembl. The column name is used to extract the statistic results, the p-values, and the FDR adjusted p-values. Then the ensembl column is used for matching in some cases. See fetch_data() for more details.

model_type
A named element of the modeling_results list. By default that is either enrichment for the model that tests one human brain layer against the rest (one group vs the rest), pairwise which compares two layers (groups) denoted by layerA-layerB such that layerA is greater than layerB, and anova which determines if any layer (group) is different from the rest adjusting for the mean expression level. The statistics for enrichment and pairwise are t-statistics while the anova model ones are F-statistics.

reverse
A logical(1) indicating whether to multiply by -1 the input statistics and reverse the layerA-layerB column names (using the -) into layerB-layerA.

Value

A table in long format with the enrichment results using stats::fisher.test().

Details

Check https://github.com/LieberInstitute/HumanPilot/blob/master/Analysis/Layer_Guesses/check_clinical_gene_sets.R to see a full script from where this family of functions is derived from.

See also

Author

Andrew E Jaffe, Leonardo Collado-Torres

Examples

## Read in the SFARI gene sets included in the package
asd_sfari <- utils::read.csv(
    system.file(
        "extdata",
        "SFARI-Gene_genes_01-03-2020release_02-04-2020export.csv",
        package = "spatialLIBD"
    ),
    as.is = TRUE
)

## Format them appropriately
asd_sfari_geneList <- list(
    Gene_SFARI_all = asd_sfari$ensembl.id,
    Gene_SFARI_high = asd_sfari$ensembl.id[asd_sfari$gene.score < 3],
    Gene_SFARI_syndromic = asd_sfari$ensembl.id[asd_sfari$syndromic == 1]
)

## Obtain the necessary data
if (!exists("modeling_results")) {
    modeling_results <- fetch_data(type = "modeling_results")
}
#> snapshotDate(): 2021-08-04
#> 2021-08-05 03:57:34 loading file /github/home/.cache/R/BiocFileCache/6c972449578f_Human_DLPFC_Visium_modeling_results.Rdata%3Fdl%3D1

## Compute the gene set enrichment results
asd_sfari_enrichment <- gene_set_enrichment(
    gene_list = asd_sfari_geneList,
    modeling_results = modeling_results,
    model_type = "enrichment"
)

## Explore the results
asd_sfari_enrichment
#>           OR         Pval   test                   ID model_type fdr_cut
#> 1  1.2659915 3.187330e-03     WM       Gene_SFARI_all enrichment     0.1
#> 2  1.1819109 1.775383e-01     WM      Gene_SFARI_high enrichment     0.1
#> 3  1.2333378 3.197106e-01     WM Gene_SFARI_syndromic enrichment     0.1
#> 4  0.9702022 8.516414e-01 Layer1       Gene_SFARI_all enrichment     0.1
#> 5  0.7192630 1.473598e-01 Layer1      Gene_SFARI_high enrichment     0.1
#> 6  1.1216176 7.379007e-01 Layer1 Gene_SFARI_syndromic enrichment     0.1
#> 7  2.7377140 5.965310e-21 Layer2       Gene_SFARI_all enrichment     0.1
#> 8  2.7066379 1.252575e-09 Layer2      Gene_SFARI_high enrichment     0.1
#> 9  2.6632367 3.564638e-04 Layer2 Gene_SFARI_syndromic enrichment     0.1
#> 10 1.3579958 2.647572e-01 Layer3       Gene_SFARI_all enrichment     0.1
#> 11 1.1738012 6.222717e-01 Layer3      Gene_SFARI_high enrichment     0.1
#> 12 2.8947133 5.518757e-02 Layer3 Gene_SFARI_syndromic enrichment     0.1
#> 13 1.2423009 2.434061e-01 Layer4       Gene_SFARI_all enrichment     0.1
#> 14 1.1445522 6.204613e-01 Layer4      Gene_SFARI_high enrichment     0.1
#> 15 2.6106289 1.575232e-02 Layer4 Gene_SFARI_syndromic enrichment     0.1
#> 16 2.0969125 8.658687e-07 Layer5       Gene_SFARI_all enrichment     0.1
#> 17 2.0956628 1.194573e-03 Layer5      Gene_SFARI_high enrichment     0.1
#> 18 0.7064982 8.016960e-01 Layer5 Gene_SFARI_syndromic enrichment     0.1
#> 19 2.6716353 1.779962e-07 Layer6       Gene_SFARI_all enrichment     0.1
#> 20 2.6206690 5.845493e-04 Layer6      Gene_SFARI_high enrichment     0.1
#> 21 2.2573853 7.927915e-02 Layer6 Gene_SFARI_syndromic enrichment     0.1